@@ -10,7 +10,7 @@ public class SecurityConfiguration {
 
     @Bean
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http.authorizeRequests()
+        http.authorizeHttpRequests()
             .anyRequest().authenticated()
             .and()
             .oauth2ResourceServer().jwt();
